requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
pyScss==1.4.0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
